requirements.txt
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
fabric==2.5.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Markdown==0.3
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
idna==2.9
invoke==1.4.1
itsdangerous==1.1.0
Jinja2==2.11.3
Markdown==3.2.2
MarkupSafe==1.1.1
mongoengine==0.20.0
oauthlib==3.1.0
paramiko==2.7.1
praw==7.0.0
prawcore==1.3.0
pycparser==2.20
pymongo==3.10.1
PyNaCl==1.3.0
PySocks==1.7.1
python-dotenv==0.13.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.8.0
update-checker==0.17
urllib3==1.25.9
uWSGI==2.0.18
websocket-client==0.57.0
Werkzeug==1.0.1
WTForms==2.3.1
